chore: update to v2.7.0 with multiple fixes and enhancements
- Support JSON5 in VS Code for enhanced development experience
- Update react-hooks to v5
- Add support for top-level await exceptions in Markdown and Vue files
- Improve import sorting with Perfectionist for cleaner codebase

+++ src/configs/yaml.ts
@@ -1,5 +1,6 @@
 import type { OptionsFiles, OptionsOverrides, OptionsStylistic, TypedFlatConfigItem } from "../types";
 import { GLOB_YAML } from "../globs";
+
 import { interopDefault } from "../utils";
 
 export async function yaml(
@@ -13,7 +14,7 @@ export async function yaml(
 
   const {
     indent = 2,
-    quotes = "double",
+    quotes = "single",
   } = typeof stylistic === "boolean" ? {} : stylistic;
 
   const [
